requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
kaleido==0.2.1
MarkupSafe==3.0.2
narwhals==1.39.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
plotly==6.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
